部分错误修正，缘分分析限制

+++ campus-love-backend/src/main/resources/db/schema.sql
@@ -354,7 +354,7 @@ CREATE TABLE IF NOT EXISTS t_yuanfen_analysis_log (
     id              BIGINT PRIMARY KEY AUTO_INCREMENT,
     user_id_a       BIGINT          NOT NULL COMMENT '较小的用户ID（保证对称性）',
     user_id_b       BIGINT          NOT NULL COMMENT '较大的用户ID',
-    total_score     TINYINT         COMMENT '触发时综合匹配分',
+    total_score     SMALLINT        COMMENT '触发时综合匹配分',
     ai_result       JSON            COMMENT 'AI返回结果快照',
     tokens_used     INT             COMMENT 'Token消耗量（成本统计）',
     created_at      DATETIME        DEFAULT CURRENT_TIMESTAMP,
